Add Windows simulator certification coverage
Add a native named-pipe command transport for the Windows all-devices app, share command dispatch with POSIX, and make metadata-driven Python tests portable to Windows direct-IP execution. Enable groupcast, automate occupancy tests, and fix the Python write-response status ABI.


Copilot-Session: db656ec6-5b4b-46bf-b0bb-a282cecd2a98

+++ src/controller/python/matter/clusters/Attribute.py
@@ -1014,7 +1014,7 @@ def _OnReadDoneCallback(closure):
 
 
 _OnWriteResponseCallbackFunct = CFUNCTYPE(
-    None, py_object, c_uint16, c_uint32, c_uint32, c_uint16)
+    None, py_object, c_uint16, c_uint32, c_uint32, c_uint8)
 _OnWriteErrorCallbackFunct = CFUNCTYPE(
     None, py_object, PyChipError)
 _OnWriteDoneCallbackFunct = CFUNCTYPE(
